third_party/WebKit/LayoutTests/typedcssom/cssPerspective.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(() => {
  let percentValue = new CSSUnitValue(10, 'percent');
  let angleValue = new CSSUnitValue(20, 'deg');
  assert_throws(new TypeError(), () => { new CSSPerspective(percentValue) });
  assert_throws(new TypeError(), () => { new CSSPerspective(angleValue) });
}, "Constructor should throw an error for invalid CSSUnitValues");

test(function() {
  var simpleLength = new CSSUnitValue(10, 'percent');
  assert_throws(new TypeError(), function() { new CSSPerspective(simpleLength) });
}, "Constructor should throw an error for CSSUnitValues with a percentage type");

test(function() {
  var perspectiveTransformSimple = new CSSPerspective(
      new CSSUnitValue(10, 'px'));
  assert_equals(perspectiveTransformSimple.toString(), 'perspective(10px)');
}, "toString should return perspective(<CSSUnitValue.cssString()>)");

</script>

<body>
</body>
